# frontend/src/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Glory RS API",
    "description": "Template API — Rust + Axum + OpenAPI",
    "contact": {
      "name": "1ndoryu"
    },
    "license": {
      "name": "MIT"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/notifications": {
      "get": {
        "tags": [
          "notifications"
        ],
        "operationId": "list_notifications",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "perPage",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "unreadOnly",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedNotifications"
                }
              }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "session_cookie": []
          }
        ]
      }
    },
    "/api/notifications/read-all": {
      "put": {
        "tags": [
          "notifications"
        ],
        "operationId": "mark_all_read",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarkAllNotificationsReadResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "session_cookie": []
          }
        ]
      }
    },
    "/api/notifications/unread-count": {
      "get": {
        "tags": [
          "notifications"
        ],
        "operationId": "unread_count",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnreadNotificationCount"
                }
              }
            }
          }
        },
        "security": [
          {
            "session_cookie": []
          }
        ]
      }
    },
    "/api/notifications/{id}": {
      "delete": {
        "tags": [
          "notifications"
        ],
        "operationId": "remove_notification",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID de notificación",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Notificación eliminada"
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "session_cookie": []
          }
        ]
      }
    },
    "/api/notifications/{id}/read": {
      "put": {
        "tags": [
          "notifications"
        ],
        "operationId": "mark_read",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID de notificación",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Notification"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "session_cookie": []
          }
        ]
      }
    },
...
    "/api/projects/{legacy_id}": {
      "put": {
        "tags": [
          "projects"
        ],
        "operationId": "upsert_project",
        "parameters": [
          {
            "name": "legacy_id",
            "in": "path",
            "description": "ID legacy del proyecto",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertProjectRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Proyecto guardado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductivityWriteResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflicto de versión",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Error de validación",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "session_cookie": []
          }
        ]
      }
    },
...
    "/api/tasks/{legacy_id}": {
      "put": {
        "tags": [
          "tasks"
        ],
        "operationId": "upsert_task",
        "parameters": [
          {
            "name": "legacy_id",
            "in": "path",
            "description": "ID legacy de la tarea",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertTaskRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Tarea guardada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductivityWriteResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflicto de versión",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Error de validación",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "session_cookie": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "description": "Estructura de respuesta de error expuesta en la API",
        "required": [
          "error",
          "message"
        ],
        "properties": {
          "error": {
            "type": "string",
            "description": "Tipo de error (`not_found`, `unauthorized`, etc.)"
          },
          "message": {
            "type": "string",
            "description": "Mensaje legible para el usuario"
          }
        }
      },
...
      "MarkAllNotificationsReadResponse": {
        "type": "object",
        "required": [
          "marked"
        ],
        "properties": {
          "marked": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "Notification": {
        "type": "object",
        "required": [
          "id",
          "notificationType",
          "title",
          "read",
          "createdAt",
          "metadata"
        ],
        "properties": {
          "content": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "metadata": {},
          "notificationType": {
            "type": "string"
          },
          "read": {
            "type": "boolean"
          },
          "readAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "title": {
            "type": "string"
          }
        }
      },
...
      "PaginatedNotifications": {
        "type": "object",
        "required": [
          "items",
          "page",
          "perPage",
          "hasMore",
          "total"
        ],
        "properties": {
          "hasMore": {
            "type": "boolean"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Notification"
            }
          },
          "page": {
            "type": "integer",
            "format": "int64"
          },
          "perPage": {
            "type": "integer",
            "format": "int64"
          },
          "total": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "ProductivityWriteResponse": {
        "type": "object",
        "required": [
          "id",
          "item",
          "updatedAt"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "item": {},
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "UnreadNotificationCount": {
        "type": "object",
        "required": [
          "unread"
        ],
        "properties": {
          "unread": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "UpsertProjectRequest": {
        "type": "object",
        "required": [
          "nombre"
        ],
        "properties": {
          "estado": {
            "type": "string"
          },
          "expectedUpdatedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "fechaLimite": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "nombre": {
            "type": "string"
          },
          "orden": {
            "type": "integer",
            "format": "int32"
          },
          "payload": {},
          "prioridad": {
            "type": "string",
            "nullable": true
          },
          "urgencia": {
            "type": "string"
          }
        }
      },
      "UpsertTaskRequest": {
        "type": "object",
        "required": [
          "texto"
        ],
        "properties": {
          "completado": {
            "type": "boolean"
          },
          "expectedUpdatedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "orden": {
            "type": "integer",
            "format": "int32"
          },
          "parentId": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "payload": {},
          "prioridad": {
            "type": "string",
            "nullable": true
          },
          "proyectoId": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "texto": {
            "type": "string"
          },
          "urgencia": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "session_cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session_id"
      }
    }
  }
}