# src/models/productivity.rs
use chrono::{DateTime, Utc};
use serde::{Deserialize, Serialize};
use serde_json::{Map, Value};
use utoipa::ToSchema;
use uuid::Uuid;

use super::dashboard::object_with_id;

fn empty_payload() -> Value {
    Value::Object(Map::new())
}

#[derive(Debug, Deserialize, ToSchema)]
#[serde(rename_all = "camelCase")]
pub struct UpsertProjectRequest {
    pub nombre: String,
    #[serde(default = "default_project_status")]
    pub estado: String,
    #[serde(default)]
    pub prioridad: Option<String>,
    #[serde(default = "default_urgency")]
    pub urgencia: String,
    #[serde(default)]
    pub fecha_limite: Option<DateTime<Utc>>,
    #[serde(default)]
    pub orden: i32,
    #[serde(default = "empty_payload")]
    pub payload: Value,
    #[serde(default)]
    pub expected_updated_at: Option<DateTime<Utc>>,
}

impl UpsertProjectRequest {
    #[must_use]
    pub fn payload_for_storage(&self, legacy_id: i64) -> Value {
        let mut object = object_with_id(self.payload.clone(), legacy_id);
        object.insert("nombre".into(), Value::String(self.nombre.clone()));
        object.insert("estado".into(), Value::String(self.estado.clone()));
        object.insert(
            "prioridad".into(),
            self.prioridad.clone().map_or(Value::Null, Value::String),
        );
        object.insert("urgencia".into(), Value::String(self.urgencia.clone()));
        object.insert(
            "fechaLimite".into(),
            serde_json::to_value(self.fecha_limite).expect("valid datetime serializes"),
        );
        object.insert("orden".into(), Value::from(self.orden));
        Value::Object(object)
    }
}

#[derive(Debug, Deserialize, ToSchema)]
#[serde(rename_all = "camelCase")]
pub struct UpsertTaskRequest {
    pub texto: String,
    #[serde(default)]
    pub completado: bool,
    #[serde(default)]
    pub prioridad: Option<String>,
    #[serde(default = "default_urgency")]
    pub urgencia: String,
    #[serde(default, rename = "proyectoId")]
    pub proyecto_id: Option<i64>,
    #[serde(default, rename = "parentId")]
    pub parent_id: Option<i64>,
    #[serde(default)]
    pub orden: i32,
    #[serde(default = "empty_payload")]
    pub payload: Value,
    #[serde(default)]
    pub expected_updated_at: Option<DateTime<Utc>>,
}

impl UpsertTaskRequest {
    #[must_use]
    pub fn payload_for_storage(&self, legacy_id: i64) -> Value {
        let mut object = object_with_id(self.payload.clone(), legacy_id);
        object.insert("texto".into(), Value::String(self.texto.clone()));
        object.insert("completado".into(), Value::Bool(self.completado));
        object.insert(
            "prioridad".into(),
            self.prioridad.clone().map_or(Value::Null, Value::String),
        );
        object.insert("urgencia".into(), Value::String(self.urgencia.clone()));
        object.insert(
            "proyectoId".into(),
            self.proyecto_id.map_or(Value::Null, Value::from),
        );
        object.insert(
            "parentId".into(),
            self.parent_id.map_or(Value::Null, Value::from),
        );
        object.insert("orden".into(), Value::from(self.orden));
        Value::Object(object)
    }

    /// Extrae la asignación nativa (`asignadoA` como UUID de Rust). El front envía la
    /// clave `asignadoA` (contrato legacy de forma) con el UUID del destinatario como
    /// valor; la columna tipada es la fuente consultable, sin correlación legacy.
    #[must_use]
    pub fn asignado_user_id(&self) -> Option<Uuid> {
        self.payload
            .get("asignadoA")
            .and_then(Value::as_str)
            .and_then(|value| Uuid::parse_str(value).ok())
    }
}

#[derive(Debug, Serialize, ToSchema)]
#[serde(rename_all = "camelCase")]
pub struct ProductivityWriteResponse {
    pub id: i64,
    pub item: Value,
    pub updated_at: DateTime<Utc>,
}

fn default_project_status() -> String {
    "activo".to_owned()
}

fn default_urgency() -> String {
    "normal".to_owned()
}

#[cfg(test)]
mod tests {
    use super::{UpsertProjectRequest, UpsertTaskRequest};
    use chrono::Utc;
    use serde_json::json;

    #[test]
    fn task_storage_payload_is_canonical_and_uses_legacy_id() {
        let request = UpsertTaskRequest {
            texto: "Nueva tarea".into(),
            completado: true,
            prioridad: Some("alta".into()),
            urgencia: "normal".into(),
            proyecto_id: Some(7),
            parent_id: None,
            orden: 2,
            payload: json!({"id": 99, "texto": "stale"}),
            expected_updated_at: None,
        };

        let value = request.payload_for_storage(11);
        assert_eq!(value["id"], json!(11));
        assert_eq!(value["texto"], json!("Nueva tarea"));
        assert_eq!(value["proyectoId"], json!(7));
    }

    #[test]
    fn task_assignment_reads_native_uuid_from_payload() {
        let uuid = "a1b2c3d4-0000-4000-8000-000000000001";
        let assigned = UpsertTaskRequest {
            texto: "Asignada".into(),
            completado: false,
            prioridad: None,
            urgencia: "normal".into(),
            proyecto_id: None,
            parent_id: None,
            orden: 0,
            payload: json!({ "asignadoA": uuid }),
            expected_updated_at: None,
        };
        assert_eq!(assigned.asignado_user_id(), Some(uuid.parse().unwrap()));

        let invalid = UpsertTaskRequest {
            texto: "Invalida".into(),
            completado: false,
            prioridad: None,
            urgencia: "normal".into(),
            proyecto_id: None,
            parent_id: None,
            orden: 0,
            payload: json!({ "asignadoA": "no-es-un-uuid" }),
            expected_updated_at: None,
        };
        assert_eq!(invalid.asignado_user_id(), None);

        let none = UpsertTaskRequest {
            texto: "Sin asignar".into(),
            completado: false,
            prioridad: None,
            urgencia: "normal".into(),
            proyecto_id: None,
            parent_id: None,
            orden: 0,
            payload: json!({}),
            expected_updated_at: None,
        };
        assert_eq!(none.asignado_user_id(), None);
    }

    #[test]
    fn project_request_preserves_conflict_timestamp() {
        let timestamp = Utc::now();
        let request = UpsertProjectRequest {
            nombre: "Glory".into(),
            estado: "activo".into(),
            prioridad: None,
            urgencia: "normal".into(),
            fecha_limite: None,
            orden: 0,
            payload: json!({}),
            expected_updated_at: Some(timestamp),
        };
        assert_eq!(request.expected_updated_at, Some(timestamp));
    }
}